Move mixin inline. It's only used in place now.
