+++ resources/js/components/fields/FieldtypeSelector.vue
@@ -74,11 +74,8 @@
 
 <script>
 import Fuse from 'fuse.js';
-import ProvidesFieldtypes from '../fields/ProvidesFieldtypes';
 
 export default {
-    mixins: [ProvidesFieldtypes],
-
     props: {
         allowTitle: {
             default: false,
@@ -128,6 +125,20 @@ export default {
     },
 
     computed: {
+        fieldtypes() {
+            if (this.fieldtypesLoading) return;
+
+            return this.$store.state.statamic.fieldtypes;
+        },
+
+        fieldtypesLoading() {
+            return this.$store.state.statamic.fieldtypes === 'loading';
+        },
+
+        fieldtypesLoaded() {
+            return Array.isArray(this.fieldtypes);
+        },
+
         allFieldtypes() {
             if (!this.fieldtypesLoaded) return [];
 
@@ -226,6 +237,18 @@ export default {
                 });
             },
         },
+    },
+
+    created() {
+        if (this.fieldtypes || this.fieldtypesLoading) return;
+
+        this.$store.commit('statamic/fieldtypes', 'loading');
+
+        let url = cp_url('fields/fieldtypes?selectable=true');
+
+        if (this.$config.get('isFormBlueprint')) url += '&forms=true';
+
+        this.$axios.get(url).then((response) => this.$store.commit('statamic/fieldtypes', response.data));
     },
 
     methods: {
